♻️ refactor: Group 도메인의 exception을 GroupAndMemberException으로 수정
