+++ src/main/java/com/grepp/spring/infra/error/exceptions/group/NotGroupLeaderException.java
@@ -1,23 +1,17 @@
 package com.grepp.spring.infra.error.exceptions.group;
 
-import com.grepp.spring.infra.response.GroupErrorCode;
+import com.grepp.spring.infra.response.GroupAndMemberErrorCode;
+import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 
+@Getter
 @Slf4j
 public class NotGroupLeaderException extends RuntimeException {
 
-    private final GroupErrorCode code;
+    private final GroupAndMemberErrorCode code;
 
-    public NotGroupLeaderException(GroupErrorCode code) {
+    public NotGroupLeaderException(GroupAndMemberErrorCode code) {
         this.code = code;
     }
 
-    public NotGroupLeaderException(GroupErrorCode code, Exception e) {
-        this.code = code;
-        log.error(e.getMessage(), e);
-    }
-
-    public GroupErrorCode code() {
-        return code;
-    }
 }
