+++ src/main/java/com/grepp/spring/infra/error/exceptions/group/NotGroupUserException.java
@@ -1,23 +1,17 @@
 package com.grepp.spring.infra.error.exceptions.group;
 
-import com.grepp.spring.infra.response.GroupErrorCode;
+import com.grepp.spring.infra.response.GroupAndMemberErrorCode;
+import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
+@Getter
 public class NotGroupUserException extends RuntimeException {
 
-    private final GroupErrorCode code;
+    private final GroupAndMemberErrorCode code;
 
-    public NotGroupUserException(GroupErrorCode code) {
+    public NotGroupUserException(GroupAndMemberErrorCode code) {
         this.code = code;
     }
 
-    public NotGroupUserException(GroupErrorCode code, Exception e) {
-        this.code = code;
-        log.error(e.getMessage(), e);
-    }
-
-    public GroupErrorCode code() {
-        return code;
-    }
 }
