+++ src/main/java/com/grepp/spring/infra/error/exceptions/group/NotScheduleLeaderException.java
@@ -1,23 +1,21 @@
 package com.grepp.spring.infra.error.exceptions.group;
 
-import com.grepp.spring.infra.response.GroupErrorCode;
+import com.grepp.spring.infra.response.GroupAndMemberErrorCode;
+import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
+@Getter
 public class NotScheduleLeaderException extends RuntimeException {
 
-    private final GroupErrorCode code;
+    private final GroupAndMemberErrorCode code;
 
-    public NotScheduleLeaderException(GroupErrorCode code) {
+    public NotScheduleLeaderException(GroupAndMemberErrorCode code) {
         this.code = code;
     }
 
-    public NotScheduleLeaderException(GroupErrorCode code, Exception e) {
+    public NotScheduleLeaderException(GroupAndMemberErrorCode code, Exception e) {
         this.code = code;
         log.error(e.getMessage(), e);
-    }
-
-    public GroupErrorCode code() {
-        return code;
     }
 }
