+++ src/main/java/com/grepp/spring/infra/error/exceptions/group/NotScheduleUserException.java
@@ -1,23 +1,21 @@
 package com.grepp.spring.infra.error.exceptions.group;
 
-import com.grepp.spring.infra.response.GroupErrorCode;
+import com.grepp.spring.infra.response.GroupAndMemberErrorCode;
+import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
+@Getter
 public class NotScheduleUserException extends RuntimeException {
 
-    private final GroupErrorCode code;
+    private final GroupAndMemberErrorCode code;
 
-    public NotScheduleUserException(GroupErrorCode code) {
+    public NotScheduleUserException(GroupAndMemberErrorCode code) {
         this.code = code;
     }
 
-    public NotScheduleUserException(GroupErrorCode code, Exception e) {
+    public NotScheduleUserException(GroupAndMemberErrorCode code, Exception e) {
         this.code = code;
         log.error(e.getMessage(), e);
-    }
-
-    public GroupErrorCode code() {
-        return code;
     }
 }
