+++ src/main/java/com/grepp/spring/infra/error/exceptions/group/OnlyOneGroupLeaderException.java
@@ -1,24 +1,16 @@
 package com.grepp.spring.infra.error.exceptions.group;
 
-import com.grepp.spring.infra.response.GroupErrorCode;
+import com.grepp.spring.infra.response.GroupAndMemberErrorCode;
+import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
+@Getter
 public class OnlyOneGroupLeaderException extends RuntimeException {
 
-    private final GroupErrorCode code;
+    private final GroupAndMemberErrorCode code;
 
-    public OnlyOneGroupLeaderException(GroupErrorCode code) {
+    public OnlyOneGroupLeaderException(GroupAndMemberErrorCode code) {
         this.code = code;
     }
-
-    public OnlyOneGroupLeaderException(GroupErrorCode code, Exception e) {
-        this.code = code;
-        log.error(e.getMessage(), e);
-    }
-
-    public GroupErrorCode code() {
-        return code;
-    }
-
 }
