+++ src/main/java/com/grepp/spring/infra/error/exceptions/group/UserNotInGroupException.java
@@ -1,24 +1,15 @@
 package com.grepp.spring.infra.error.exceptions.group;
 
-import com.grepp.spring.infra.response.GroupErrorCode;
+import com.grepp.spring.infra.response.GroupAndMemberErrorCode;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
 public class UserNotInGroupException extends RuntimeException {
 
-    private final GroupErrorCode code;
+    private final GroupAndMemberErrorCode code;
 
-    public UserNotInGroupException(GroupErrorCode code) {
+    public UserNotInGroupException(GroupAndMemberErrorCode code) {
         this.code = code;
-    }
-
-    public UserNotInGroupException(GroupErrorCode code, Exception e) {
-        this.code = code;
-        log.error(e.getMessage(), e);
-    }
-
-    public GroupErrorCode code() {
-        return code;
     }
 
 }
